# docs/swagger.yaml
basePath: /api
definitions:
  handlers.arusKasRequest:
    properties:
      id_bank:
        example: 1
        type: integer
      jenis:
        enum:
        - Pemasukan
        - Pengeluaran
        example: Pengeluaran
        type: string
      kategori:
        example: Biaya Operasional
        type: string
      keterangan:
        example: Pembelian ATK
        type: string
      nominal:
        example: 500000
        type: number
      tanggal:
        example: "2026-04-14"
        type: string
    required:
    - jenis
    - kategori
    - nominal
    type: object
...
  services.BookingInput:
    properties:
      id_customer:
        type: integer
      id_kavling:
        type: integer
      id_marketing:
        type: integer
      id_user:
        type: integer
      keterangan:
        type: string
      nominal_booking:
        type: number
      tgl_expired:
        type: string
    required:
    - id_customer
    - id_kavling
    type: object
  services.CashInput:
    properties:
      harga_jual:
        type: number
      id_customer:
        type: integer
      id_kavling:
        type: integer
      id_marketing:
        type: integer
      id_user:
        type: integer
      keterangan:
        type: string
    required:
    - harga_jual
    - id_customer
    - id_kavling
    type: object
  services.ConvertInput:
    properties:
      harga_jual:
        type: number
      id_user:
        type: integer
      jenis:
        description: cash | kredit
        type: string
      keterangan:
        type: string
      lama_cicilan:
        type: integer
      tgl_mulai_cicilan:
        type: string
      uang_muka:
        type: number
    required:
    - harga_jual
    - jenis
    type: object
  services.KreditInput:
    properties:
      harga_jual:
        type: number
      id_customer:
        type: integer
      id_kavling:
        type: integer
      id_marketing:
        type: integer
      id_user:
        type: integer
      keterangan:
        type: string
      lama_cicilan:
        minimum: 1
        type: integer
      tgl_mulai_cicilan:
        type: string
      uang_muka:
        type: number
    required:
    - harga_jual
    - id_customer
    - id_kavling
    - lama_cicilan
    type: object
host: localhost:8080
info:
  contact:
    email: admin@kavling-mentaya.com
    name: API Support
  description: REST API backend untuk sistem manajemen penjualan kavling tanah. Semua
    endpoint (kecuali /auth/login) memerlukan Bearer Token JWT di header Authorization.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Kavling Mentaya API
  version: "1.0"
paths:
...
  /booking:
    get:
      description: Mengambil daftar semua booking yang masih aktif (status=1)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List booking aktif
      tags:
      - Booking
    post:
      consumes:
      - application/json
      description: Membuat booking kavling. Kavling harus berstatus Kosong (0). Status
        kavling berubah menjadi Booking (1).
      parameters:
      - description: Data booking
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.BookingInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Buat booking baru
      tags:
      - Booking
  /booking/{id}:
    delete:
      description: Membatalkan booking aktif. Status kavling dikembalikan ke Kosong
        (0).
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Batalkan booking
      tags:
      - Booking
  /booking/{id}/convert:
    post:
      consumes:
      - application/json
      description: Mengubah booking menjadi transaksi cash atau kredit. Jika kredit,
        cicilan per bulan otomatis dihitung = (harga_jual - uang_muka) / lama_cicilan.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Data konversi (jenis: cash|kredit)'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.ConvertInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Konversi booking ke transaksi
      tags:
      - Booking
...
  /keuangan/rekap-kredit:
    get:
      description: 'Menghitung rekap cicilan semua kavling kredit: bulan berjalan,
        jumlah pembayaran, tunggakan, nominal tunggakan, dan status bulan ini.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Rekap kredit semua kavling
      tags:
      - Keuangan
  /keuangan/statistik:
    get:
      description: Mengambil rekap pemasukan dan pengeluaran per bulan untuk tahun
        tertentu
      parameters:
      - description: 'Tahun (default: tahun sekarang)'
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Statistik arus kas per bulan
      tags:
      - Keuangan
  /keuangan/transaksi:
    get:
      description: Mengambil daftar transaksi arus kas dengan filter tanggal. Response
        juga mencakup total pemasukan, pengeluaran, dan saldo.
      parameters:
      - description: Dari tanggal (YYYY-MM-DD)
        in: query
        name: dari
        type: string
      - description: Sampai tanggal (YYYY-MM-DD)
        in: query
        name: sampai
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List arus kas
      tags:
      - Keuangan
    post:
      consumes:
      - application/json
      description: Mencatat pemasukan atau pengeluaran secara manual (bukan dari cicilan).
        Saldo bank otomatis diupdate.
      parameters:
      - description: Data transaksi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.arusKasRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tambah transaksi arus kas manual
      tags:
      - Keuangan
  /keuangan/transaksi/{id}:
    delete:
      description: Menghapus transaksi arus kas manual. Saldo bank otomatis dikembalikan.
      parameters:
      - description: Transaksi ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Hapus transaksi arus kas
      tags:
      - Keuangan
...
  /pembayaran:
    get:
      description: Mengambil semua record pembayaran cicilan beserta info kavling
        dan customer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List semua pembayaran cicilan
      tags:
      - Pembayaran
  /pembayaran/{id_kavling}:
    get:
      description: Mengambil riwayat pembayaran cicilan untuk kavling tertentu beserta
        data transaksinya
      parameters:
      - description: Kavling ID
        in: path
        name: id_kavling
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Detail pembayaran per kavling
      tags:
      - Pembayaran
  /pembayaran/{id_kavling}/bayar:
    post:
      consumes:
      - multipart/form-data
      description: Mencatat pembayaran cicilan untuk kavling kredit. Otomatis mencatat
        ke arus kas. Jika sudah mencapai lama_cicilan, status kavling berubah menjadi
        CASH/Lunas.
      parameters:
      - description: Kavling ID
        in: path
        name: id_kavling
        required: true
        type: integer
      - description: Jumlah pembayaran
        in: formData
        name: jumlah_bayar
        required: true
        type: number
      - description: ID rekening bank tujuan
        in: formData
        name: id_bank
        type: integer
      - description: 'Tanggal bayar (YYYY-MM-DD, default: hari ini)'
        in: formData
        name: tanggal
        type: string
      - description: Keterangan
        in: formData
        name: keterangan
        type: string
      - description: Bukti transfer (gambar)
        in: formData
        name: bukti_pembayaran
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Bayar cicilan
      tags:
      - Pembayaran
  /transaksi:
    get:
      description: Mengambil daftar semua transaksi jual-beli kavling (cash dan kredit)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List semua transaksi kavling
      tags:
      - Transaksi Kavling
  /transaksi/{id}:
    get:
      description: Mengambil detail satu transaksi kavling berdasarkan ID
      parameters:
      - description: Transaksi ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Detail transaksi kavling
      tags:
      - Transaksi Kavling
  /transaksi/cash:
    post:
      consumes:
      - application/json
      description: Membuat transaksi pembelian cash langsung dari peta kavling (tanpa
        booking). Kavling harus berstatus Kosong (0).
      parameters:
      - description: Data transaksi cash
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CashInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Transaksi cash langsung
      tags:
      - Transaksi Kavling
  /transaksi/kredit:
    post:
      consumes:
      - application/json
      description: Membuat transaksi pembelian kredit langsung dari peta kavling.
        Cicilan per bulan = (harga_jual - uang_muka) / lama_cicilan. Tgl jatuh tempo
        default tanggal 10 bulan berikutnya.
      parameters:
      - description: Data transaksi kredit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.KreditInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Transaksi kredit langsung
      tags:
      - Transaksi Kavling
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}. Token diperoleh
      dari endpoint POST /api/auth/login.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"